kompost

Module anonymous

Structs

AnonymousIterator

Iterator that with a context and a closures called in the AnonymousIterator::new and AnonymousIterator::next methods.

Traits

Anonymous

Trait that adds a method to Iterator structs (i.e., Iterators, slices, and Vecs) that allows painlessly create anonynmous iterators for one-time use.