kompost

Module composite

Collection of composite method that useful (to me)

Functions

circular_windows

A composite function to be used with the crate::Composed::composed method that takes an additional single usize as a parameter and computes a window of that size for every element of the iterator (circular, it takes elements from the beginning for later windows).

circular_windows_2d

composite function to generate circular sliding windows over a 2D data structure in form of an Iterator over Iterator. See [this example](crate#complex-example section) for how to use it.

circular_windows_2d_slice

composite function to generate circular sliding windows over a 2D data structure in form of an Iterator over slices (such as returned by the chunks method) See [this example](crate#complex-example section) for how to use it.

transpose

TBD: Transpose over Iterable of Iterable

transpose_slice

Function to be used with the crate::Composed::composed method. It transposes an IntoIterator of std::slice, a data structure often encountered when storing 2D arrays in a single (row-major) array and using chunks for iteration