kompost

Module compounds

Functions

periodic_windows

A compound function to be used with the crate::ComposedIterable::composed method that takes an additional single usize as a parameter and computes a window of that size for every element of the iterator (periodic).

transpose

Function to be used with the crate::ComposedIterable::composed method. It transposes an IntoIterator of std::slice, a data structure often encountered when storing 2D arrays in a single (row-major) array.