kompost::compounds

Function circular_windows_2d

pub fn circular_windows_2d<T>(
    it: impl ExactSizeIterator<Item = impl ExactSizeIterator<Item = T> + Clone> + Clone,
    size_m: usize,
    size_n: usize,
) -> impl Iterator<Item = impl Iterator<Item = impl Iterator<Item = impl Iterator<Item = T>>>>

Compound function to generate circular sliding windows over a 2D data structure in form of an Iterator over Iterator. See [this example](crate#complex-example section) for how to use it.